requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.4
pydantic==2.9.2
redis==5.2.0
openai==1.58.1
sentence-transformers==3.2.1
python-dotenv==1.0.1
alembic==1.13.3
